+++ addon/utils/position-utils.ts
@@ -1,1 +1,2 @@
 export * from './_private/position-utils';
+export * from './_private/document-walker';